soniq-agent/requirements.txt
livekit-agents>=1.4.3
livekit-plugins-deepgram>=1.0.0
livekit-plugins-cartesia>=1.0.0
livekit-plugins-openai>=1.0.0
livekit-plugins-silero>=1.0.0
livekit-plugins-noise-cancellation>=0.2.0
livekit-plugins-turn-detector>=1.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
